Implement shrinking map over time
Commits adds the functionality for the map to shrink over time. Current
limits are set such that one square will be filled in for sure each turn
past the 500 turn mark. Also now saving the current map state in the
MAPSTATE variable, as using the static map from the start would not
suffice to visualize the game.